+++ src/Exceptions/FailureToBuildStep.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace MeadSteve\Tale\Exceptions;
+
+class FailureToBuildStep extends \RuntimeException implements TaleException
+{
+
+}