+++ src/Exceptions/TaleException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace MeadSteve\Tale\Exceptions;
+
+interface TaleException extends \Throwable
+{
+
+}