+++ CHANGELOG.md
@@ -5,10 +5,12 @@
 ### Enhancements
 * Add finalising Step interface - useful for committing transactions etc.
 * Catch all Errors not just exceptions.
+* Add builder to help with constructing steps via calls to ::add(...$args)
 
 ### Bug fixes
 
 ### Backwards incompatible changes
+* ::addStep() made private
 
 
 ## v0.1.0 (2018-06-19)